requirements.txt
djangorestframework
pillow
Django
channels
channels-redis
redis
gunicorn
daphne